Organize nf modules and workflows;  integrate genomefactory
