+++ nextflow/README.md
@@ -1,0 +1,6 @@
+### Clone Required Repositories
+ensembl-production
+ensembl-metadata-api
+```
+git clone ensembl-production
+```
