+++ nextflow/nextflow.config
@@ -1,0 +1,93 @@
+
+
+// default params
+params.pipeline_dir = "$PWD"
+params.user = "$USER"
+params.email = "${params.user}@ebi.ac.uk"
+params.email_server = "hh-smtp.ebi.ac.uk:25"
+params.base_dir = "$BASE_DIR"
+params.scripts_dir = "${params.base_dir}/nf-sql-to-parquet/"
+params.query_dir = "${params.scripts_dir}/queries"
+params.output_dir = "${NOBACKUP_DIR}/parquet/"
+
+//production nf configs
+prams.ensembl_production_nf_base = "${param.base_dir}/ensembl-production/nextflow/"
+prams.ensembl_production_nf_modules = "${ensembl_production_nf_base}/modules/"
+params.ensembl_production_nf_scripts = "${prams.ensembl_production_nf_base}/nf-py-scripts/"
+
+
+// genome factory default params
+genome_uuid           = ""
+dataset_uuid          = ""
+organism_group_type   = ""
+division              = ""
+dataset_type          = "genebuild"
+species               = ""
+antispecies           = ""
+dataset_status        = "Submitted"
+update_dataset_status = ""
+batch_size            = "50"
+page                  = ""
+columns               = ""
+output_json           = "genome_info.json"
+params.core_db_host_uri = "mysql://anonymous@ensembldb.ensembl.org/",
+params.metadata_db_uri = "mysql://anonymous@ensembldb.ensembl.org/ensembl_genome_metadata"
+
+
+//default directives
+profile = "slurm"
+cleanup = true
+workDir = "${params.output_dir}/work"
+publishDir = "${params.output_dir}"
+
+trace {
+    enabled = true
+    file = "trace"
+}
+
+report {
+    file = "report.html"
+    enable = true
+}
+
+profiles {
+
+  slurm {
+
+   process {
+
+    errorStrategy = { task.attempt <= process.maxRetries  ? 'retry' : 'finish' }
+
+    executor = 'slurm'
+    queue = 'production'
+    queueSize = 100
+    maxRetries = 3
+    time = '1d'
+
+    withLabel:small_process {
+        memory = 200.MB
+    }
+
+    withLabel: dm {
+        queue = 'datamover'
+        time = '5h'
+        memory = 2.GB
+    }
+    withLabel:mem2GB {
+        memory = 2.GB
+
+    }
+    withLabel:mem4GB {
+        memory = 4.GB
+
+    }
+  }
+
+ }
+
+}
+
+
+
+
+
