--- src/ensembl/production/nextflow/nextflow.config
@@ -1,10 +1,0 @@
-// Config directory with all the queries
-params.query_dir = "queries"
-
-// Config directory with all the species
-params.species_dir = "species"
-
-// output directory
-params.target_dir = "multi_species"
-
-cleanup = true
