+++ src/ensembl/production/sql_to_parquet/queries/species/btaurus.json
