+++ src/ensembl/production/sql_to_parquet/queries/species/mmusculus.json
